ms_sms_management/applications/app-service/src/test/java/co.com.bancolombia.config/UseCasesConfigTest.java
package co.com.bancolombia.config;

import co.com.bancolombia.model.log.Log;
import co.com.bancolombia.model.log.gateways.LogGateway;
import co.com.bancolombia.model.message.Alert;
import co.com.bancolombia.model.message.Response;
import co.com.bancolombia.model.message.TemplateSms;
import co.com.bancolombia.usecase.log.LogUseCase;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import reactor.core.publisher.Mono;
import reactor.test.StepVerifier;

import java.util.UUID;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class UseCasesConfigTest {
    @InjectMocks
    private UseCasesConfig useCasesConfig;


    @BeforeEach
    public void init (){

    }

    @Test
    public void putLogTest(){
        assertThat(useCasesConfig.customOpenAPI("1")).isNotNull();
    }

}
